+++ UVtools.Core/Extensions/PointExtensions.cs
@@ -9,13 +9,14 @@ namespace UVtools.Core.Extensions
         
         public static Point Rotate(this Point point, double angleDegree, Point pivot = default)
         {
+            if (angleDegree == 0 || angleDegree == 360) return point;
             double angle = angleDegree * Math.PI / 180;
             double cos = Math.Cos(angle);
             double sin = Math.Sin(angle);
             int dx = point.X - pivot.X;
             int dy = point.Y - pivot.Y;
             double x = cos * dx - sin * dy + pivot.X;
-            double y = sin * dx + cos * dy + pivot.X;
+            double y = sin * dx + cos * dy + pivot.Y;
 
             Point rotated = new Point((int)Math.Round(x), (int)Math.Round(y));
             return rotated;